+++ documentation/rebuild_docs_and_distribute.py
@@ -32,11 +32,11 @@ CAMCOPS_ROOT_DIR = os.path.abspath(os.path.join(THIS_DIR, os.pardir))
 BUILD_HTML_DIR = os.path.join(THIS_DIR, "build", "html")
 SERVER_DOCS_DIR = os.path.join(CAMCOPS_ROOT_DIR, "server", "camcops_server",
                                "static", "documentation_copy")
-WEBSITE_DOCS_DIR = os.path.join(CAMCOPS_ROOT_DIR, "website", "documentation")
+# WEBSITE_DOCS_DIR = os.path.join(CAMCOPS_ROOT_DIR, "website", "documentation")
 
 DEST_DIRS = [
     SERVER_DOCS_DIR,
-    WEBSITE_DOCS_DIR,
+    # WEBSITE_DOCS_DIR,
 ]
 
 # Remove anything old